Refactor - move utils out of image annotation field view model
